Update EBMS reports so can be used for UKBMS
The EBMS Sites/Sections download reports had a conditional to ensure the
relevant location had no parent. This is on top of a check on the
location_type_id. Unfortunately some of the Sites in UKBMS are assigned
to branches using the parent_id, so they did not appear on the report.
Given that the location_type_id is used as well, this change just
removes the parent_id check.